+++ docs/guides/other_libs/samples/ModifyLogMethod.cs
@@ -6,8 +6,8 @@ private static async Task LogAsync(LogMessage message)
         LogSeverity.Error => LogEventLevel.Error,
         LogSeverity.Warning => LogEventLevel.Warning,
         LogSeverity.Info => LogEventLevel.Information,
-        LogSeverity.Verbose => LogEventLevel.Verbose,
-        LogSeverity.Debug => LogEventLevel.Debug,
+        LogSeverity.Verbose => LogEventLevel.Debug,
+        LogSeverity.Debug => LogEventLevel.Verbose,
         _ => LogEventLevel.Information
     };
     Log.Write(severity, message.Exception, "[{Source}] {Message}", message.Source, message.Message);