+++ src/Discord.Net.Core/Entities/Messages/IAttachment.cs
@@ -62,5 +62,13 @@ namespace Discord
         ///     <see langword="true"/> if the attachment is ephemeral; otherwise <see langword="false"/>.
         /// </returns>
         bool Ephemeral { get; }
+        /// <summary>
+        ///     Gets the description of the attachment; or <see langword="null"/> if there is none set.
+        /// </summary>
+        string Description { get; }
+        /// <summary>
+        ///     Gets the media's <see href="https://en.wikipedia.org/wiki/Media_type">MIME type</see> if present; otherwise <see langword="null"/>.
+        /// </summary>
+        string ContentType { get; }
     }
 }