+++ src/Discord.Net.WebSocket/DiscordSocketClient.cs
@@ -1311,7 +1311,7 @@ namespace Discord.WebSocket
                                         else
                                         {
                                             user = guild.AddOrUpdateUser(data);
-                                            var cacheableBefore = new Cacheable<SocketGuildUser, ulong>(user, user.Id, true, () => null);
+                                            var cacheableBefore = new Cacheable<SocketGuildUser, ulong>(null, user.Id, false, () => null);
                                             await TimedInvokeAsync(_guildMemberUpdatedEvent, nameof(GuildMemberUpdated), cacheableBefore, user).ConfigureAwait(false);
                                         }
                                     }
@@ -2017,6 +2017,92 @@ namespace Discord.WebSocket
                                 break;
                             #endregion
 
+                            #region Integrations
+                            case "INTEGRATION_CREATE":
+                                {
+                                    await _gatewayLogger.DebugAsync("Received Dispatch (INTEGRATION_CREATE)").ConfigureAwait(false);
+
+                                    var data = (payload as JToken).ToObject<Integration>(_serializer);
+
+                                    // Integrations from Gateway should always have guild IDs specified.
+                                    if (!data.GuildId.IsSpecified)
+                                        return;
+
+                                    var guild = State.GetGuild(data.GuildId.Value);
+
+                                    if (guild != null)
+                                    {
+                                        if (!guild.IsSynced)
+                                        {
+                                            await UnsyncedGuildAsync(type, guild.Id).ConfigureAwait(false);
+                                            return;
+                                        }
+
+                                        await TimedInvokeAsync(_integrationCreated, nameof(IntegrationCreated), RestIntegration.Create(this, guild, data)).ConfigureAwait(false);
+                                    }
+                                    else
+                                    {
+                                        await UnknownGuildAsync(type, data.GuildId.Value).ConfigureAwait(false);
+                                        return;
+                                    }
+                                }
+                                break;
+                            case "INTEGRATION_UPDATE":
+                                {
+                                    await _gatewayLogger.DebugAsync("Received Dispatch (INTEGRATION_UPDATE)").ConfigureAwait(false);
+
+                                    var data = (payload as JToken).ToObject<Integration>(_serializer);
+
+                                    // Integrations from Gateway should always have guild IDs specified.
+                                    if (!data.GuildId.IsSpecified)
+                                        return;
+
+                                    var guild = State.GetGuild(data.GuildId.Value);
+
+                                    if (guild != null)
+                                    {
+                                        if (!guild.IsSynced)
+                                        {
+                                            await UnsyncedGuildAsync(type, guild.Id).ConfigureAwait(false);
+                                            return;
+                                        }
+
+                                        await TimedInvokeAsync(_integrationUpdated, nameof(IntegrationUpdated), RestIntegration.Create(this, guild, data)).ConfigureAwait(false);
+                                    }
+                                    else
+                                    {
+                                        await UnknownGuildAsync(type, data.GuildId.Value).ConfigureAwait(false);
+                                        return;
+                                    }
+                                }
+                                break;
+                            case "INTEGRATION_DELETE":
+                                {
+                                    await _gatewayLogger.DebugAsync("Received Dispatch (INTEGRATION_DELETE)").ConfigureAwait(false);
+
+                                    var data = (payload as JToken).ToObject<IntegrationDeletedEvent>(_serializer);
+
+                                    var guild = State.GetGuild(data.GuildId);
+
+                                    if (guild != null)
+                                    {
+                                        if (!guild.IsSynced)
+                                        {
+                                            await UnsyncedGuildAsync(type, guild.Id).ConfigureAwait(false);
+                                            return;
+                                        }
+
+                                        await TimedInvokeAsync(_integrationDeleted, nameof(IntegrationDeleted), guild, data.Id, data.ApplicationID).ConfigureAwait(false);
+                                    }
+                                    else
+                                    {
+                                        await UnknownGuildAsync(type, data.GuildId).ConfigureAwait(false);
+                                        return;
+                                    }
+                                }
+                                break;
+                            #endregion
+
                             #region Users
                             case "USER_UPDATE":
                                 {
@@ -2245,7 +2331,7 @@ namespace Discord.WebSocket
 
                                     SocketUser user = data.User.IsSpecified
                                         ? State.GetOrAddUser(data.User.Value.Id, (_) => SocketGlobalUser.Create(this, State, data.User.Value))
-                                        : guild.AddOrUpdateUser(data.Member.Value);
+                                        : guild?.AddOrUpdateUser(data.Member.Value); // null if the bot scope isn't set, so the guild cannot be retrieved.
 
                                     SocketChannel channel = null;
                                     if(data.ChannelId.IsSpecified)
@@ -2260,8 +2346,12 @@ namespace Discord.WebSocket
                                             }
                                             else
                                             {
-                                                await UnknownChannelAsync(type, data.ChannelId.Value).ConfigureAwait(false);
-                                                return;
+                                                if (guild != null) // The guild id is set, but the guild cannot be found as the bot scope is not set.
+                                                {
+                                                    await UnknownChannelAsync(type, data.ChannelId.Value).ConfigureAwait(false);
+                                                    return;
+                                                }
+                                                // The channel isnt required when responding to an interaction, so we can leave the channel null.
                                             }
                                         }
                                     }